Req guia 06

+++ src/main/java/isi/dan/laboratorios/danmsusuarios/service/ClienteServiceImp.java
@@ -1,7 +1,7 @@
 package isi.dan.laboratorios.danmsusuarios.service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import isi.dan.laboratorios.danmsusuarios.domain.Cliente;
 import isi.dan.laboratorios.danmsusuarios.exception.RecursoNoEncontradoException;
 import isi.dan.laboratorios.danmsusuarios.exception.RiesgoException;
-import isi.dan.laboratorios.danmsusuarios.repository.ClienteRepository;
+import isi.dan.laboratorios.danmsusuarios.dao.ClienteRepository;
 
 @Service
 public class ClienteServiceImp implements ClienteService {
@@ -42,18 +42,20 @@ public class ClienteServiceImp implements ClienteService {
 
     @Override
     public void darBaja(Integer id) throws RecursoNoEncontradoException {
-        try {
-            Cliente cliente = clienteRepository.findById(id).get();
+        
+            Optional<Cliente> response = clienteRepository.findById(id);
+
+            if(response.isEmpty()) throw new RecursoNoEncontradoException();
+
+            Cliente cliente = response.get();
 
             if (!pedidoService.pedidosDelCliente(cliente).isEmpty()) {
-                cliente.setFechaBaja(new Date());
+                cliente.setFechaBaja(LocalDateTime.now());
                 clienteRepository.save(cliente);
             } else {
                 clienteRepository.delete(cliente);
             }
-        } catch (Exception e) {
-            throw new RecursoNoEncontradoException();
-        }
+        
     }
 
     @Override
@@ -69,6 +71,7 @@ public class ClienteServiceImp implements ClienteService {
         if (cuit != null) {
             clientes.removeIf(cliente -> !cuit.equals(cliente.getCuit()));
         }
+
         if (razon != null) {
             clientes.removeIf(cliente -> !razon.equals(cliente.getRazonSocial()));
         }
@@ -78,7 +81,7 @@ public class ClienteServiceImp implements ClienteService {
 
     @Override
     public Cliente buscarPorId(Integer id) throws RecursoNoEncontradoException {
-        Optional<Cliente> cliente = clienteRepository.findById(id).filter(c -> c.getFechaBaja() == null);
+        Optional<Cliente> cliente = clienteRepository.findByIdAndFechaBajaIsNull(id);
 
         if (cliente.isPresent())
             return cliente.get();
@@ -87,7 +90,7 @@ public class ClienteServiceImp implements ClienteService {
 
     @Override
     public Cliente actualizar(Cliente nuevo) throws RecursoNoEncontradoException {
-        Optional<Cliente> cliente = clienteRepository.findById(nuevo.getId());
+        Optional<Cliente> cliente = clienteRepository.findByIdAndFechaBajaIsNull(nuevo.getId());
 
         if (cliente.isPresent()) {
             return clienteRepository.save(nuevo);
